requirements.txt
torch
torchvision
gradio
Pillow
fpdf2
